Add contexts to show
When running `pd show ###` the output now includes the attached
contexts (links attached when the alert was created).

As part of this change, the Alert class was reworked a little, mostly to
move the body into its own class, instead of having the pd cli
manipulate the raw object or force additional details into the Alert
class.

Though now that I'm thinking about it, it might've just made sense to
get rid of the `alert.body` property, and just add contexts and details
directly to the alert, because the only thing we really care about on it
is the body... Something to consider at least.